+++ engine/src/analyzer/report/report.c
@@ -17,25 +17,37 @@ vmcu_report_t* vmcu_report_ctor(const char *file) {
     report->progsize    = 0;
     report->disassembly = NULL;
 
-    report->n_label = 0;
-    report->label   = NULL;
+    report->n_sfr       = 0;
+    report->sfr         = NULL;
+
+    report->n_label     = 0;
+    report->label       = NULL;
 
     return report; 
 }
 
 void vmcu_report_dtor(vmcu_report_t *this) {
     
-    for(int i = 0; i < this->progsize; i++) {
+    for(int32_t i = 0; i < this->progsize; i++) {
 
         if(this->disassembly[i].mnem != NULL)
             free(this->disassembly[i].mnem);
     }
 
-    for(int i = 0; i < this->n_label; i++) {
+    for(int32_t i = 0; i < this->n_sfr; i++) {
+
+        if(this->sfr[i].n_xref > 0)
+            free(this->sfr[i].xref);
+    }
+
+    for(int32_t i = 0; i < this->n_label; i++) {
 
         if(this->label[i].n_xref > 0)
             free(this->label[i].xref);
     }
+
+    if(this->sfr != NULL)
+        free(this->sfr);
 
     if(this->label != NULL)
         free(this->label);