+++ engine/src/system/core/sfreg.c
@@ -1,7 +1,7 @@
 /* SFR Implementation */
 
 // Project Headers (engine)
-#include "engine/include/system/core/sfr.h"
+#include "engine/include/system/core/sfreg.h"
 #include "engine/include/system/core/io.h"
 #include "engine/include/system/core/irq.h"
 #include "engine/include/system/peripherals/timer8.h"